style: address light DOM UI
